requirements.txt
pystray
Pillow
requests
